backend/docker/requirements.txt
azure-iothub-device-client==1.4.0
azure-iothub-service-client==1.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.3
click==6.7
configparser==3.5.0
decorator==4.3.0
entrypoints==0.2.3
enum34==1.1.6
Flask==1.0.2
futures==3.2.0
html5lib==1.0.1
ipykernel==4.8.2
ipython==5.7.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
pandocfilters==1.4.2
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pyzmq==17.0.0
qtconsole==4.3.1
scandir==1.7
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1